requirements.txt
Django==6.0.4
django-filter==25.2
djangorestframework==3.17.1